mcp-servers/aws-api/requirements.txt
mcp==1.26.0
awslabs.aws-api-mcp-server==1.3.9
fastmcp==2.14.5
boto3==1.42.47
pydantic==2.12.5